chore: change layout
